Revert "Feat: 요약 및 문제 생성 API 구현 (#34)"
This reverts commit 0c85c1aae73e1a32e1147ec35caa10113acd6e57.

+++ src/main/java/notai/member/application/MemberService.java
@@ -16,7 +16,6 @@ public class MemberService {
     public Long login(Member member) {
         return memberRepository.findByOauthId(member.getOauthId())
                 .orElseGet(() -> memberRepository.save(member))
-                .getId()
-                .longValue();
+                .getId();
     }
 }
